remove script
